health_scraper/requirements.txt
scrapy
requests
openai
upstash_vector
python-dotenv
langchain
uuid